Final refactoring - Deliver price
